+++ physics/CHANGELOG.md
@@ -8,5 +8,9 @@ changes will be documented in this file.
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+- Fixed panic in wire baseline calibration for new runs by removing "safeguards".
+
 <!-- next-url -->
 [Unreleased]: https://github.com/ALPHA-g-Experiment/alpha-g/compare/alpha_g_physics-v0.1.2...HEAD